+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Public/SpanOptions.cs
@@ -7,5 +7,6 @@ namespace BugsnagUnityPerformance
         public ISpanContext ParentContext = null;
         public bool MakeCurrentContext = true;
         public bool? IsFirstClass = null;
+        public bool? InstrumentRendering = null;
     }
 }