+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/NestedSpans/NewThreadNewContext.cs
@@ -3,27 +3,33 @@ using System.Collections.Generic;
 using BugsnagUnityPerformance;
 using System.Threading;
 using UnityEngine;
-
+using UnityEngine.Scripting;
 public class NewThreadNewContext : Scenario
 {
+    [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.NoInlining)]
     public override void PreparePerformanceConfig(string apiKey, string host)
     {
         base.PreparePerformanceConfig(apiKey, host);
         SetMaxBatchSize(3);
     }
-
+    [System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.NoInlining)]
     public override void Run()
     {
+        StartCoroutine(DoRun());
+    }
+
+    // Unity 2021 IL2CPP keeps messing up the threads for some reason, so we stick them in a coroutine because it leaves them alone
+    private IEnumerator DoRun()
+    {
         var span1 = BugsnagPerformance.StartSpan("span1");
-        var finished = false;
-        new Thread(() => {
+        new Thread(() =>
+        {
             var span2 = BugsnagPerformance.StartSpan("span2");
-                var span3 = BugsnagPerformance.StartSpan("span3");
-                span3.End();
+            var span3 = BugsnagPerformance.StartSpan("span3");
+            span3.End();
             span2.End();
-            finished = true;
         }).Start();
-        while (!finished) { }
+        yield return new WaitForSeconds(1);
         span1.End();
     }
 }