+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v1.8.0 (2025-02-05)
+
+### Additions
+
+- Add frame rate metrics to spans. [#155](https://github.com/bugsnag/bugsnag-unity-performance/pull/155)
+
+### Bug Fixes
+
+- Fix an issue where spans started with null names sent no name field. Now string.Empty will be sent instead [#154](https://github.com/bugsnag/bugsnag-unity-performance/pull/154)
+
 ## v1.7.1 (2024-12-06)
 
 ### Bug Fixes